requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
djangorestframework==3.14.0
drf-spectacular==0.25.1
flake8==6.0.0
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
